Start Tax Crud

+++ src/main/java/com/luxrest/rm/Tax/TaxRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaxRepository extends JpaRepository<Tax, Long> {
+public interface TaxRepository extends JpaRepository<Tax, Integer> {
 }
